Remove stray comments
